fix GCC -Wall/-Wextra/-Wpedantic compiler warnings
fixed bugs:
- bullsquid not reacting to being attacked while hunting a headcrab
- attempt to play a null sentence
- wrong reload anim for body guard when using mp5
- a mystery bug with the base grunt AI (probably something rare)
- maybe some random crashes caused by violating strict aliasing

+++ dlls/Bsp.cpp
@@ -46,7 +46,7 @@ bool Bsp::load_lumps(string fpath)
 	int size = fin.tellg();
 	fin.seekg(0, fin.beg);
 
-	if (size < sizeof(BSPHEADER) + sizeof(BSPLUMP) * HEADER_LUMPS) {
+	if (size < (int)(sizeof(BSPHEADER) + sizeof(BSPLUMP) * HEADER_LUMPS)) {
 		update_lump_pointers();
 		valid = false;
 		return false;
